feat: move hack to shared crate
This commit gives up on const fn for parsing strings, instead we move
Hack instructions to a shared crate and use strum::EnumString.

This implementation is not `const` but in a future commit we will
introduce a `proc_macro` that will parse all the literals and inject
`Instruction` literals in their place.